Comments added to my code. File rework based on comments from TA. I also attempted at running tests for my use cases but not successful. I kept the work available to show the attempt at the tests.
