+++ src/main/java/APP_Business_Rules/create_user/CreateUserResponseModel.java
@@ -4,7 +4,7 @@ public class CreateUserResponseModel {
 
     String login;
 
-
+    //Optional message loginmsg can be stored for testing purposes.
     public CreateUserResponseModel(String loginmsg) {
         this.login = loginmsg;
     }
