+++ src/main/java/APP_Business_Rules/login_user/LoginUserResponseModel.java
@@ -4,7 +4,10 @@ public class LoginUserResponseModel {
 
     LoginUserGatewayModel model;
 
-
+    /**
+     * The result for when a user has attempted a login.
+     * @param model the information found from a login.
+     */
     public LoginUserResponseModel(LoginUserGatewayModel model) {
         this.model = model;
     }
@@ -12,4 +15,6 @@ public class LoginUserResponseModel {
     public String getUsername(){
         return this.model.getUsername();
     }
+
+
 }
