+++ src/main/java/Interface_and_Adapters/WelcomeScreen.java
@@ -1,4 +1,4 @@
-package Interface_and_Adapters.StartUpScreens;
+package Interface_and_Adapters;
 
 import APP_Business_Rules.login_user.LoginUserResponseModel;
