+++ src/main/java/Interface_and_Adapters/start_up_screens/LabelHelper.java
@@ -1,4 +1,4 @@
-package Interface_and_Adapters.StartUpScreens;
+package Interface_and_Adapters.start_up_screens;
 
 import javax.swing.*;
