+++ src/main/java/Interface_and_Adapters/start_up_screens/SignUpScreen.java
@@ -1,6 +1,5 @@
-package Interface_and_Adapters.StartUpScreens;
+package Interface_and_Adapters.start_up_screens;
 
-import APP_Business_Rules.create_user.CreateUserController;
 import APP_Business_Rules.login_user.*;
 import Entities.AccountFactory;
 import Entities.UserFactory;
